Change tidy resource to a cron job to clean old files
Prior to this commit, the module used a tidy resource to clean
old metrics files which only works if the module is applied
regularly but not in a puppet apply once and leave it scenario.

After this commit, we move the tidy resource into a cron job
that runs puppet apply.  This allows users to puppet apply the
module and not run into unbounded growth of the output directory