Add ensure parameter and add hosts parameters to init.pp
Prior to this commit, you could only set the hosts for metrics
via hiera data.  If you enabled metrics and wanted to disable
them you could only do so manually.

After this commit, hosts for each metric are parameters in init.pp.
There is a metric_ensure parameter that can be set to absent
to remove the scripts and cron jobs.  The output directory will
always be made and not removed.