Swap out puppet apply tidy for find commands in cron
When using puppet apply with the tidy resource we
found that users with large amounts of data were
wasting a lot of resources deleting the old files
because tidy is not very efficient.

This commit changes over to using the find command
which should be available on all PE supported
platforms.