Bump version to 2.0.0
The 2.0.0 release is a major release because we remove the
tidy_age parameter from the pe_metric defined type and replaced it
with the retention days parameter.  Since this parameter was not
exposed in init.pp we suspect that this change will not affect
many users but it is still semantically a breaking change.