Implement outputdir/host/timestamp.json structure
This commit updates the scripts to create one directory per server
metrics are being obtained from, and to place files for each server in
its dedicated directory.

The idea is to make it easier to manually browse records on the
filesystem using cd and ls, for very large sets of data.