Adhere to Unix philosophy "Rule of Silence"
The cleanup script which runs nightly runs `find [...] -delete -print`
to delete old reports. The -print flag causes the find command to
produce output like the following on stdout - which is then emailed to
users.

    /opt/puppetlabs/pe_metric_curl_cron_jobs/puppetserver/127.0.0.1/20170304T025501Z.json
    /opt/puppetlabs/pe_metric_curl_cron_jobs/puppetdb/127.0.0.1/20170304T025501Z.json

This is a normal recurring activity and does not need to be logged.
Sending emails for it violates the "Rule of Silence" in the Unix
philosophy. When a program has nothing surprising to say, it should say
nothing.

This commit aligns the retention cleanup cron job with the Unix
philosophy's rule of silence.