Use new metric for command queue depth in PE 2017.1
Prior to this commit, we used the activemq metrics to determine
the command queue depth.

After this commit, when the PE version is 2071.1 or higher and
thus includes the new stockpile queue implementation we use the
new metric for determining the queue depth.  Otherwise we continue
to use the amq metrics.