Cleanup old scripts after renaming with .sh extension
Prior to this commit, if you upgrade from a < 2 version of the
module to a > 2 version of the module you'd see that you had a
puppetdb_metrics.sh script and a puppet_server_metrics.sh script
in your scripts directory.

After this commit, we ensure that those scripts are removed.