Pull the PuppetDB status endpoint for consistent queue_depth metric
Prior to this commit, we used ActiveMQ metrics to find the queue
depth metric and then in 2017.1 we used a specific metric from
PuppetDB.  However, the status endpoint has a queue_depth metric
in it that correctly chooses the backend depending on version.

After this commit, we will only track the status endpoint and not
the specific queue depth metric so that we can look at the same
metric across versions.