Allow passing a different metric script template to pe_metric
Prior to this commit, we consolidated the puppetdb and puppetserver
metric script templates to be the same reducing complexity.
However, after doing that we wanted to add a seperate script for
tracking activemq metrics that doesn't follow the same template.

After this commit, we allow new metrics like ActiveMQ to pass in a
different script template to use.  This allows consolidation of the
puppetdb and puppetserver scripts and still allows for a one-off
metrics template for new metrics.

This commit also renames the metrics template to tk_metrics to
indicate it is made for interacting with trapper keeper metrics.