Include hour, minute, and second in compressed tarball name
Prior to this commit, running the script to compress metrics files
twice in one day would result in overwriting the existing tarball.
If you want to run the script manually for debugging or any other
reason, losing data would be a not preferred side effect.

After this commit, the tarball name includes the hour, minute and
second to avoid overwriting the tarball if run twice in one day.