Compress per host instead of across all hosts
Prior to this commit, it was very difficult to use
bzgrep to grep for metrics because you couldn't tell
which host they came from.

After this commit, we compress the metrics per host
which allows a user to use bzgrep in the directory
of the host and find metrics for just that host.