Move metrics hash out of the template and into puppet code
Prior to this commit, we had a metrics hash for puppetdb in the
tk_metrics template.  Since puppetserver and orchestrator use that
template it wasn't the right place for the metrics hash.

After this commit, the metrics hash is passed in via a parameter
to the template and PuppetDB figures out the hash to pass in via
its own metrics class.