Retrieve all metrics in one POST instead of multiple GETs
Prior to this commit, we ran one GET request to the metrics
endpoint per metric needed.

After this commit, we POST an array of metrics to retrieve from
the endpoint and get them all in one request.

This is documented here:
https://docs.puppet.com/puppetdb/4.4/api/metrics/v1/mbeans.html#bulk-retrieving-metrics

As a side effect, we no longer url encode the metrics in the
metrics hash that is stored in puppet code.