Add a --print commandline argument to the tk_metrics script
Prior to this commit, the tk_metrics script only wrote its
output to a file.

After this commit, you can pass a cmdline argument to have the
output written to stdout in addition to its output file.  This
allows for integrations with other tools that can read the
output from stdout.