Load script configuration from a yaml file
Prior to this commit, all the variables that needed to be passed
from Puppet code into the ruby script for collecting metrics were
injected via an epp template.

After this commit, the configuration variables will be passed into
the script from a yaml file.  This commit adds a new argument to
the script which is the metrics_type so the correct configuration
file can be found.