Add a --no-file argument to allow disabling storing a file of metrics
Prior to this commit, whenever the script ran it always saved output
to a file.  You could additionally use the --print argument to get the
output on STDOUT for use by another program or script.  However, if you
wanted to run the script often, say every 10 seconds, you'd produce a
lot of files and potentially use too much disk space.

After this commit, you can run the script with `--no-file` and the
script will not save the output to the file.  This is intended to be
used only by a setup outside of the cron jobs provided by this module.
The cron jobs should always save output on their regular schedule but
we want to provide an option to not store the file for use cases that
run the script a lot via another mechanism.