Implement --output-dir argument
Prior to this commit, you could only specify the output_dir from
the config file for the metric type you were calling.

After this commit, you can specify the output_dir from the CLI.

In addition, if you do not specify an --output-dir then output
will not be written to a file.