Fix retention script to delete really old files
Prior to this commit, the find command to delete old files would
only delete files exactly $retention_days older than when the
command is run.  This means that if the script runs every day then
everything should be fine and no extra data is piling up.  However,
if the server is off or the cron does not run for any reason then
extra data will be left behind as it will be older than the
retention_days parameter.

After this commit, the find command will delete files older than
the retention days parameter instead of exactly as old as the
retention_days parameter.

We also switch to ctime instead of mtime beacuse we care about when
the files were created not when they were last modified.