No longer pass certificate information to netHTTP
Prior to this commit, we set the cacert, cert, and key but also
pass SSL_VERIFY_NONE which negates the need for the former.

After this commit, we do not set the cacert, cert, and key so that
non-root users can run the script without hitting permissions
errors on the keys.