URI escape metrics when retriveing them individually
In a past commit we stopped saving the metric urls already
escaped and moved to POSTing all of the metrics at that time.

Then we realized we needed to still GET the metrics individually
for older PE versions but we didn't consider they'd need to be
re-escaped.

This commit adds back escaping when metrics are retrieved
individually.