Connect over http by default when puppetdb on loopback
Prior to this commit, we'd connect to puppetdb over https and
not verify certs.  This doesn't work prior to 2016.4 because
puppetdb has metrics behind a cert whitelist.

After this commit, we connect to puppetdb over http if it is
running on the same node as the metrics script.  This avoids
the certificate issue.

For PE versions prior to 2016.4. you will not be able to collect
metrics on remote puppetdb systems.