Improve support for configuration file
Make a new option, --config, which is used to directly specify the
configuration file to use. Also clean up the variable names so the code
is easier to follow in terms of what it's doing.

The old behavior where --metrics-type implied a config file name is
still supported as legacy.