Remove support for metrics-type legacy config file
The script must now be invoked with the --config option to specify a
configuration file, OR all required configuration parameters must be
passed as command-line flags.

Because this involves a change in the API to interact with the script,
the amq_metrics script is now in scope also, and has had the same option
parsing interface applied to it.