Make unsupplied optional values clearer
Previously some optional values had a default value of False. In order
to indicate that the user doesn't need to supply a value a default is
required, but False as the default did not make it clear that the value
simply hadn't been supplied.

This commit switches such values to use :undef as their default instead,
thus making it clearer what is happening.