Add the ability to ship metrics directly to influxdb
Prior to this commit, there was no built-in way to ship metrics
to an outside metrics server.

After this commit, we pull in a script from
puppetlabs/puppet-metrics-viewer that allows conversion of the
metrics data we gather to influxdb or graphite format.

This commit adds the initial support to enabling shipping of metrics
directly to influxdb when given a hostname to connect to.